add users

+++ controllers/contactControllers.js
@@ -2,15 +2,19 @@ const asyncHandler = require('express-async-handler');
 const Contact = require('../models/contactModel');
 //@desc get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
+    if(!contacts){
+        res.status(404);
+        throw new Error("Contacts not found");
+    }       
     res.status(200).json(contacts);
 });
 //@desc create all contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
     console.log("the request body:", req.body);
@@ -22,31 +26,40 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
 });
 
 // @ desc delete contact
 // @ route DELETE /api/contacts/:id
-// @ access public
+// @ access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to delete other user's contact");
+    }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 })
 //@desc update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contact");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -57,13 +70,17 @@ const updateContact = asyncHandler(async (req, res) => {
 })
 //@desc get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to access other user's contact");
+    }   
     res.status(200).json(contact);
 })
 module.exports = { getContacts, createContact, deleteContact, updateContact, getContact };
